+++ src/modules/horizontal_controller.cpp
@@ -24,5 +24,5 @@ void HorizontalController::control(float x_r, float y_r, float x, float y, float
 //Control acceleration given reference position (m) and current position (m) and velocity (m/s) with given controller gains
 float HorizontalController::control_siso(float pos_r, float pos, float vel, float kp, float kd)
 {
-    return (kp * (pos_r - pos) + kd * (0 - vel));
+    return (float(kp * (pos_r - pos) + kd * (0 - vel)));
 }
